+++ Nbuild/Cli.cs
@@ -4,7 +4,7 @@ namespace Nbuild;
 
 public class Cli
 {
-    [OptionalArgument("", "cmd", "command. value = [targets]\n" +
+    [OptionalArgument("", "c", "command. value = [targets]\n" +
         "\t targets \t -> list targets in nbuild.targets and common.targets files")]
     public string? Command { get; set; }
