+++ Ngit/Program.cs
@@ -5,16 +5,17 @@ using NbuildTasks;
 using Ngit;
 using OutputColorizer;
 using System.Diagnostics;
+using System.Reflection;
 using static NbuildTasks.Enums;
 
 
-var watch = Stopwatch.StartNew();
 RetCode ReturnCode = RetCode.InvalidParameter;
+string NgitAssemblyExe = "Ngit";
 
 if (!Parser.TryParse(args, out Cli options))
 {
     if (!args[0].Equals("--help", StringComparison.CurrentCultureIgnoreCase))
-        Console.WriteLine($"Ngit completed with '-1'");
+        Console.WriteLine($"{NgitAssemblyExe} Completed with '-1'");
     return 0;
 }
 
@@ -27,11 +28,10 @@ if (!string.IsNullOrEmpty(options.GitCommand))
     ReturnCode = Command.Process(options);
 }
 
-watch.Stop();
-
 if (ReturnCode != RetCode.Success)
 {
-    Colorizer.WriteLine($"Ngit completed in {watch.ElapsedMilliseconds} ms with [{ConsoleColor.Red}!'{ReturnCode}']");
+    
+    Colorizer.WriteLine($"{NgitAssemblyExe} Completed with [{ConsoleColor.Red}!'{ReturnCode}']");
 }
 
 return Convert.ToInt32(ReturnCode);
